% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperfunc.R
\name{hypdist}
\alias{hypdist}
\title{Hyperbolic distance in the Poincaré disk}
\usage{
hypdist(u, v)
}
\arguments{
\item{u}{Numeric vector of length 2. Must satisfy \code{||u|| < 1}.}

\item{v}{Numeric vector of length 2. Must satisfy \code{||v|| < 1}.}
}
\value{
A numeric value giving the hyperbolic distance between \code{u} and \code{v}.
}
\description{
Computes the hyperbolic distance between two points inside the
2-dimensional Poincaré disk model.
}
\examples{
hypdist(c(0.1, 0), c(0.2, 0.3))

}
