% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbolize.R
\name{hyperbolize}
\alias{hyperbolize}
\title{Hyperbolic cluster mapping}
\usage{
hyperbolize(D, numk, preprocess, scale = c("std", "norm"), use_input_order = F)
}
\arguments{
\item{D}{Numeric matrix, with observations in rows and variables in columns.}

\item{numk}{Integer. Number of clusters.}

\item{preprocess}{Either a character string specifying the clustering method
(\code{"kmeansWrap"}, \code{"kmeans"}, \code{"kmedians"},
\code{"hddc"}, \code{"pgmm"}), or directly a clustering result / label vector:}

\item{scale}{Character. Scaling choice for the data:
\itemize{
  \item \code{"std"}: standardize using \code{scale(D)}.
  \item \code{"norm"}: center and rescale to a fixed average norm (paper sec 3.2).
}}

\item{use_input_order}{Logical. If \code{FALSE} (default), cluster centers are
reordered using \code{order_clusters_tsp}. If \code{TRUE}, the input cluster order is kept.}
}
\value{
A list with components:
\describe{
  \item{points}{\code{n × 2} matrix. Final 2D hyperbolic embedding of all samples.}
  \item{centers}{\code{numk × 2} matrix. 2D positions of cluster centroids.}
  \item{cluster}{Integer vector of length \code{n}. Cluster label of each sample.}
  \item{analysis}{Model fit object returned by the chosen clustering method
   , or \code{NULL} if cluster labels were supplied directly.}
}
}
\description{
This function takes a data matrix, maps both cluster centers and observations into the
2D hyperbolic disk based on given cluster labels.
}
\examples{
\dontrun{
  set.seed(100000)
  D <- scale(iris[, 1:4])
  out <- hyperbolize(D, numk = 3,scale="std", preprocess = "kmeansWrap")
  plot_disk(out,3)
}
}
